Make some arguments of __init__ mandatory and check format string
The directory and format arguments must now be specified expliitly, and
a ValueError is now raised if the format string does not contain
'{timestamp}' as a substring.

Also add docstrings for DFHist and DFHist.__init__.

Test plan: New unit tests check that a TypeError is raised if the
required arguments are not provided, and that a ValueError is raised if
the format doesn't have a substring '{timestamp}'.